check processInstance state before delete worker group

+++ dolphinscheduler-api/src/main/java/org/apache/dolphinscheduler/api/service/WorkerGroupService.java
@@ -19,8 +19,12 @@ package org.apache.dolphinscheduler.api.service;
 import org.apache.dolphinscheduler.api.enums.Status;
 import org.apache.dolphinscheduler.api.utils.Constants;
 import org.apache.dolphinscheduler.api.utils.PageInfo;
+import org.apache.dolphinscheduler.common.enums.ExecutionStatus;
+import org.apache.dolphinscheduler.common.utils.CollectionUtils;
+import org.apache.dolphinscheduler.dao.entity.ProcessInstance;
 import org.apache.dolphinscheduler.dao.entity.User;
 import org.apache.dolphinscheduler.dao.entity.WorkerGroup;
+import org.apache.dolphinscheduler.dao.mapper.ProcessInstanceMapper;
 import org.apache.dolphinscheduler.dao.mapper.WorkerGroupMapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
@@ -42,6 +46,9 @@ public class WorkerGroupService extends BaseService {
 
     @Autowired
     WorkerGroupMapper workerGroupMapper;
+
+    @Autowired
+    ProcessInstanceMapper processInstanceMapper;
 
     /**
      * create or update a worker group
@@ -142,7 +149,23 @@ public class WorkerGroupService extends BaseService {
 
         Map<String, Object> result = new HashMap<>(5);
 
+        int[] states = new int[]{
+                ExecutionStatus.SUBMITTED_SUCCESS.ordinal(),
+                ExecutionStatus.RUNNING_EXEUTION.ordinal(),
+                ExecutionStatus.READY_PAUSE.ordinal(),
+                ExecutionStatus.READY_STOP.ordinal(),
+                ExecutionStatus.NEED_FAULT_TOLERANCE.ordinal(),
+                ExecutionStatus.WAITTING_THREAD.ordinal(),
+                ExecutionStatus.WAITTING_DEPEND.ordinal()
+        };
+
+        List<ProcessInstance> processInstances = processInstanceMapper.queryByWorkerGroupIdAndStatus(id, states);
+        if(CollectionUtils.isNotEmpty(processInstances)){
+            putMsg(result, Status.DELETE_WORKER_GROUP_BY_ID_FAIL, processInstances.size());
+            return result;
+        }
         workerGroupMapper.deleteById(id);
+        processInstanceMapper.updateProcessInstanceByWorkerGroupId(id, org.apache.dolphinscheduler.common.Constants.DEFAULT_WORKER_ID);
         putMsg(result, Status.SUCCESS);
         return result;
     }
